[WIP] feat(reversibility): content tree - warning generator backend
